app/requirements/base.txt
Django>=3.2,<3.3
wagtail>=2.13,<2.14

# Third Party

# Models
django-model-utils==4.1.1
python_dateutil==2.7.5

# Utils
django-ipware==2.1.0

# DRF
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
django-allauth==0.44.0
djangorestframework-jwt==1.11.0

# Static files
whitenoise==5.2.0  # https://github.com/evansd/whitenoise


# Document
drf-yasg==1.20.0

# Celery
celery==4.2.1
eventlet==0.24.1
redis==3.2.1
django-celery-beat==1.4.0
django-celery-results==1.0.4

# Storages
boto3==1.9.182
django-storages==1.11.1

#Coverage
coverage==4.5.3

pyfcm==1.4.7
PyJWT==1.7.1

django-environ==0.4.5

# This is used for JWT
djangorestframework-simplejwt==4.6.0

# Config for redoc
PyYAML==5.4.1
uritemplate==3.0.1